extract the bingo grid to be its own widget
